+++ start.sh
@@ -1,1 +1,1 @@
-python -m SimpleHTTPServer
+python3 -m http.server 8000